models: u64_wrapper!: impl std::ops for u64/usize

+++ src/sentry2/mod.rs
@@ -5,5 +5,5 @@ pub mod types;
 
 pub use coordinator::*;
 
-pub const CHUNK_SIZE: u64 = 1 << 10;
-pub const BATCH_SIZE: u64 = 3 << 15;
+pub const CHUNK_SIZE: usize = 1 << 10;
+pub const BATCH_SIZE: usize = 3 << 15;
